==> 2.logisticreg.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

# Read the train and the test data 
#The preprocessing file needs to be run first, it creates the csv file used here
train_users = pd.read_csv('train_users_merge_scale.csv')
#test_users = pd.read_csv('test_users.csv')


# Extracting labels from the train data
train_users_labels = train_users.loc[:,'country_destination']
print (train_users_labels.head(n=5))

# Extracting attributes from the train data
train_users_attrs = train_users.iloc[:,:-1]
print(train_users_attrs.head(n=5))

train_users = train_users_attrs
print(train_users.columns)
train_users = train_users.drop(['Blackberry'], axis=1)
train_users = train_users.drop(['Opera Phone'], axis=1)
labels_df = pd.DataFrame(train_users_labels)

from sklearn.preprocessing import LabelBinarizer
from sklearn.metrics import make_scorer

def dcg_score(y_true, y_score, k=5):

    order = np.argsort(y_score)[::-1]
    y_true = np.take(y_true, order[:k])

    gain = 2 ** y_true - 1

    discounts = np.log2(np.arange(len(y_true)) + 2)
    return np.sum(gain / discounts)


#def ndcg_score(ground_truth, predictions, k=5):
def ndcg_score(te_labels, predict, k):
   
    lb = LabelBinarizer()
    lb.fit(range(12 + 1))
    T = lb.transform(te_labels)

    scores = []

    # Iterate over each y_true and compute the DCG score
    for y_true, y_score in zip(T, predict):
        actual = dcg_score(y_true, y_score, k)
        best = dcg_score(y_true, y_true, k)
        if best == 0:
            best = 0.000000001
        score = float(actual) / float(best)
        scores.append(score)
    return np.mean(scores)


# NDCG Scorer function
ndcg_scorer = make_scorer(ndcg_score, needs_proba=True, k=5)


def folds_to_split(data,targets,train,test):
    data_tr = pd.DataFrame(data).iloc[train]
    data_te = pd.DataFrame(data).iloc[test]
    labels_tr = pd.DataFrame(targets).iloc[train]
    labels_te = pd.DataFrame(targets).iloc[test]
    return [data_tr, data_te, labels_tr, labels_te]

# Logistic Regression : One Versus Rest
# Use validation set to find which solver to use 

from sklearn.multiclass import OneVsRestClassifier
from sklearn import linear_model, cross_validation

fold_results_ovr = pd.DataFrame()
test_sets = []
test_set_labels = []
train_sets = []
train_set_labels = []
    
def ten_fold_oneVsRest(data, labels):
    
    for train, test in cross_validation.KFold(len(data), n_folds=10, shuffle=True, random_state=20160202):
        [tr_data, te_data,
         tr_target, te_target] = folds_to_split(data, labels,train,test)
        
        train_sets.append(tr_data)
        train_set_labels.append(tr_target)
        test_sets.append(te_data)
        test_set_labels.append(te_target)
        
    foldnum = 0
    solvers = ['newton-cg', 'lbfgs', 'liblinear']
    for x in range(0, 10):
        foldnum+=1
        [bnb_train, bnb_validation, bnb_train_labels, bnb_validation_labels] = cross_validation.train_test_split(
        train_sets[x], 
        train_set_labels[x], 
        test_size=0.11, 
        random_state=20160121)
        
        for sol in solvers:
            oneVsRest = OneVsRestClassifier(linear_model.LogisticRegression(solver=sol))
            #print(tr_target)
            oneVsRest.fit(bnb_train.values, bnb_train_labels[bnb_train_labels.columns.values[0]].values)
            #print(oneVsRest.estimators_)
        
            fold_results_ovr.loc[foldnum, 'Accuracy_' + sol] = oneVsRest.score(bnb_validation, bnb_validation_labels)
        
            predictions = oneVsRest.predict_proba(bnb_validation)
            score = ndcg_score(bnb_validation_labels.as_matrix(), predictions, 5)
            fold_results_ovr.loc[foldnum, 'ndcg_' + sol]  = score
            #print(score)
                
    #Now let's look at the results:
    print ("------OneVsRestClassifier accuracy and ndcg values------")
    print (fold_results_ovr)
    print ("------OneVsRestClassifier mean accuracy, ndcg values------")
    print(fold_results_ovr.mean())
    
ten_fold_oneVsRest(train_users, labels_df)

fold_results_ovr = pd.DataFrame()
foldnum = 0
for x in range(0, 10):
    foldnum+=1
    oneVsRest = OneVsRestClassifier(linear_model.LogisticRegression(solver='lbfgs', penalty='l2'))
    oneVsRest.fit(train_sets[x].values, train_set_labels[x].values.ravel())
    
    fold_results_ovr.loc[foldnum, 'Accuracy'] = oneVsOne.score(test_sets[x].values,
                                                          test_set_labels[x].values.ravel())
    
    predictions = oneVsRest.predict_proba(test_sets[x].values)
    score = ndcg_score(test_set_labels[x].as_matrix(), predictions, 5)
    fold_results_ovr.loc[foldnum, 'ndcg']  = score
    
print ("------OneVsRestClassifier accuracy and ndcg values------")
print (fold_results_ovr)
print ("------OneVsRestClassifier mean accuracy, ndcg values------")
print(fold_results_ovr.mean())

# liblinear and lbfgs perform  nearly the same...

fold_results_ovr = pd.DataFrame()
foldnum = 0
for x in range(0, 10):
    foldnum+=1
    oneVsRest = OneVsRestClassifier(linear_model.LogisticRegression(solver='liblinear', penalty='l2'))
    oneVsRest.fit(train_sets[x].values, train_set_labels[x].values.ravel())
    
    fold_results_ovr.loc[foldnum, 'Accuracy'] = oneVsOne.score(test_sets[x].values,
                                                          test_set_labels[x].values.ravel())
    
    predictions = oneVsRest.predict_proba(test_sets[x].values)
    score = ndcg_score(test_set_labels[x].as_matrix(), predictions, 5)
    fold_results_ovr.loc[foldnum, 'ndcg']  = score
    
print ("------OneVsRestClassifier accuracy and ndcg values------")
print (fold_results_ovr)
print ("------OneVsRestClassifier mean accuracy, ndcg values------")
print(fold_results_ovr.mean())

# liblinear does better with l1 penalty...

fold_results_ovr = pd.DataFrame()
foldnum = 0
for x in range(0, 10):
    foldnum+=1
    oneVsRest = OneVsRestClassifier(linear_model.LogisticRegression(solver='liblinear', penalty='l1'))
    oneVsRest.fit(train_sets[x].values, train_set_labels[x].values.ravel())
    
    fold_results_ovr.loc[foldnum, 'Accuracy'] = oneVsOne.score(test_sets[x].values,
                                                          test_set_labels[x].values.ravel())
    
    predictions = oneVsRest.predict_proba(test_sets[x].values)
    score = ndcg_score(test_set_labels[x].as_matrix(), predictions, 5)
    fold_results_ovr.loc[foldnum, 'ndcg']  = score
    
print ("------OneVsRestClassifier accuracy and ndcg values------")
print (fold_results_ovr)
print ("------OneVsRestClassifier mean accuracy, ndcg values------")
print(fold_results_ovr.mean())

# Logistic Regression : One Versus One
# Use validation set to find which solver to use 

from sklearn.multiclass import OneVsOneClassifier
fold_results_ovo = pd.DataFrame()
test_sets = []
test_set_labels = []
train_sets = []
train_set_labels = []

def ten_fold_oneVsOne(data, labels):
    foldnum = 0
    solvers = ['newton-cg', 'lbfgs', 'liblinear']
    
    for train, test in cross_validation.KFold(len(data), n_folds=10, shuffle=True):
        [tr_data, te_data,
         tr_target, te_target] = folds_to_split(data, labels,train,test)
        
        train_sets.append(tr_data)
        train_set_labels.append(tr_target)
        test_sets.append(te_data)
        test_set_labels.append(te_target)
    
    foldnum = 0
    for x in range(0, 10):
        foldnum+=1
        [bnb_train, bnb_validation, bnb_train_labels, bnb_validation_labels] = cross_validation.train_test_split(
        train_sets[x], 
        train_set_labels[x], 
        test_size=0.11, 
        random_state=20160121)
        for sol in solvers:
            oneVsOne = OneVsOneClassifier(linear_model.LogisticRegression(solver=sol))
            oneVsOne.fit(bnb_train.values, bnb_train_labels[bnb_train_labels.columns.values[0]].values)
        
            columnname = "Accuracy" + "_" + sol
            fold_results_ovo.loc[foldnum, columnname] = oneVsOne.score(bnb_validation, bnb_validation_labels)
    
    #Now let's look at the results:
    print ("-----------OneVsOneClassifier results for accuracy-----------")
    print (fold_results_ovo)
    
    print ("--------------OneVsOneClassifier mean accuracy----------------")
    print(fold_results_ovo.mean())
    
ten_fold_oneVsOne(train_users, labels_df)

# lbfgs does best every time...

fold_results_ovo = pd.DataFrame()
foldnum = 0
for x in range(0, 10):
    
    foldnum+=1
    oneVsOne = OneVsOneClassifier(linear_model.LogisticRegression(solver='lbfgs'))
    oneVsOne.fit(train_sets[x].values, train_set_labels[x].values.ravel())
    
    fold_results_ovo.loc[foldnum, 'Accuracy'] = oneVsOne.score(test_sets[x].values,
                                                          test_set_labels[x].values.ravel())
    

print ("-----------OneVsOneClassifier results for accuracy-----------")
print (fold_results_ovo)
print ("--------------OneVsOneClassifier mean accuracy----------------")
print(fold_results_ovo.mean())

# Use ADA boost with logistic regression

from sklearn.ensemble import AdaBoostClassifier
from sklearn.ensemble import RandomForestClassifier

from sklearn.multiclass import OneVsRestClassifier
from sklearn import linear_model, cross_validation

test_sets = []
test_set_labels = []
train_sets = []
train_set_labels = []
fold_results_ada = pd.DataFrame()
def ten_fold_ada_logistic(data, labels):
    foldnum = 0
    fold_results_ovr_est = pd.DataFrame()
    for train, test in cross_validation.KFold(len(data), n_folds=10, shuffle=True, random_state=20160202):
        
        [tr_data, te_data,
         tr_target, te_target] = folds_to_split(data, labels,train,test)
        train_sets.append(tr_data)
        train_set_labels.append(tr_target)
        test_sets.append(te_data)
        test_set_labels.append(te_target)
    
    solvers = ['newton-cg', 'lbfgs']
    estimators = [100, 200, 250]
    foldnum = 0
    for x in range(0, 10):
        foldnum+=1
        [bnb_train, bnb_validation, bnb_train_labels, bnb_validation_labels] = cross_validation.train_test_split(
            train_sets[x], 
            train_set_labels[x], 
            test_size=0.11, 
            random_state=20160121)
        for sol in solvers:
            for est in estimators:
                
                [tr_data, te_data,
                 tr_target, te_target] = folds_to_split(data, labels,train,test)
                ada = AdaBoostClassifier(learning_rate=0.3, n_estimators=est, 
                                     base_estimator=linear_model.LogisticRegression(solver=sol), 
                                     algorithm='SAMME')  
                #print(tr_target)
                ada.fit(bnb_train.values, bnb_train_labels[bnb_train_labels.columns.values[0]].values)
                #print(oneVsRest.estimators_)
        
                fold_results_ada.loc[foldnum, 'Accuracy_' + sol + '_' + str(est)] = ada.score(bnb_validation, 
                                                                                            bnb_validation_labels)
        
                predictions = ada.predict_proba(bnb_validation)
                score = ndcg_score(bnb_validation_labels.as_matrix(), predictions, 5)
                fold_results_ada.loc[foldnum, 'ndcg_' + sol + '_' + str(est)]  = score
                print(str(score) + ":" + " solver " + sol + ", estimators " + str(est))
            
    
    #Now let's look at the results:
    print ("------Ada accuracy and ndcg values------")
    print (fold_results_ada)
    print ("------Ada mean accuracy and ndcg values------")
    print(fold_results_ada.mean())
    
ten_fold_ada_logistic(train_users, labels_df)

# SVM : Not completing, takes forever to run

from sklearn import svm
    
def ten_fold_svm(data, labels):
    foldnum = 0
    fold_results_ovr_est = pd.DataFrame()
    fold_results_ada = pd.DataFrame()
    for train, test in cross_validation.KFold(len(data), n_folds=10, shuffle=True, random_state=20160202):
        foldnum+=1
        [tr_data, te_data,
         tr_target, te_target] = folds_to_split(data, labels,train,test)
        vec_mach = svm.SVC(kernel='rbf', random_state=20160202)
        #print(tr_target)
        vec_mach.fit(tr_data.values, tr_target[tr_target.columns.values[0]].values)
        
        fold_results_ada.loc[foldnum, 'Accuracy'] = vec_mach.score(te_data, te_target)
        
        predictions = vec_mach.predict_proba(te_data)
        score = ndcg_score(te_target.as_matrix(), predictions, 5)
        fold_results_ada.loc[foldnum, 'ndcg']  = score
        print(score)
            
    
    #Now let's look at the results:
    print ("------svm accuracy values------")
    print (fold_results_ovr)
    print ("------svm mean accuracy values------")
    print(fold_results_ovr.mean())
    
ten_fold_svm(train_users, labels_df)

The best performance with logistic regression works with liblinear with l1 penalty 
 ndcg score is 81.5417
