==> 4.RandomForest.ipynb <==
# Random Forest 

This notebook tries different models on the preprocessed airbnb dataset using Random Forest Classifier.

## Part 1 : Preprocessed data with 'data_first_booking' 

import pandas as pd
import numpy as np
from sklearn import preprocessing
from sklearn import cross_validation
from sklearn.ensemble import RandomForestClassifier
from sklearn.preprocessing import LabelBinarizer
from sklearn.metrics import make_scorer

data = pd.read_csv('preprocessed_airbnb_train.csv')
labels = data.loc[:,'country_destination']
data = data.drop(['country_destination'], axis=1)

def folds_to_split(data,targets,train,test):
    data_tr = pd.DataFrame(data).iloc[train]
    data_te = pd.DataFrame(data).iloc[test]
    labels_tr = pd.DataFrame(targets).iloc[train]
    labels_te = pd.DataFrame(targets).iloc[test]
    return [data_tr, data_te, labels_tr, labels_te]

def dcg_score(y_true, y_score, k=5):
    order = np.argsort(y_score)[::-1]
    y_true = np.take(y_true, order[:k])

    gain = 2 ** y_true - 1

    discounts = np.log2(np.arange(len(y_true)) + 2)
    return np.sum(gain / discounts)

def ndcg_score(te_labels, predict, k):
    lb = LabelBinarizer()
    lb.fit(range(len(predict) + 1))
    T = lb.transform(te_labels)
    scores = []

    for y_true, y_score in zip(T, predict):
        actual = dcg_score(y_true, y_score, k)
        best = dcg_score(y_true, y_true, k)  
        score = float(actual) / float(best)
        scores.append(score)

    return np.mean(scores)

ndcg_scorer = make_scorer(ndcg_score, needs_proba=True, k=5)


### Varying the max-depth of the Random Forest 

results = pd.DataFrame()
foldnum=0

for train, test in cross_validation.KFold(len(data), shuffle=True, n_folds=10,
                                           random_state=20160217):
    foldnum+=1
    [tr_data, te_data, tr_labels, te_labels] = folds_to_split(data, labels, train, test)

    for max_depth in range(1, 5):
        rf = RandomForestClassifier(n_estimators=600,criterion='gini', max_depth=max_depth)
        rf = rf.fit(tr_data, tr_labels.values.ravel())
        rf_predict = rf.predict_proba(te_data) 
        score = ndcg_score(te_labels.as_matrix(), rf_predict, k=5)
        print 'Fold : {}, Max-depth : {}, Score : {}'.format( foldnum, max_depth, score )
        results.loc[foldnum, 'max_depth=%d'%max_depth ] = score

results.mean()

It can be seen that the nDCG score increases with the max-depth.

### Without setting the max-depth 

foldnum=0
results = pd.DataFrame()
for train, test in cross_validation.KFold(len(data), shuffle=True, n_folds=10,
                                           random_state=20160217):
    foldnum+=1
    [tr_data, te_data, tr_labels, te_labels] = folds_to_split(data, labels, train, test)
    rf = RandomForestClassifier(n_estimators=600,criterion='gini')
    rf = rf.fit(tr_data, tr_labels.values.ravel())
    rf_predict = rf.predict_proba(te_data) 
    score = ndcg_score(te_labels.as_matrix(), rf_predict, k=5)
    results.loc[foldnum, 'no max-depth' ] = score

results.mean()

The nDCG score of this model is comparable to max-depth = 4. This great score is attributed to the bias by the 'date_first_booking' which will be NULL for all the new users whose prediction need to be made. A good model will be the one without the 'date_first_booking' attribute. 

## Part 2 : Preprocessed data without 'date_first_booking'

As seen during the exploratory analysis, 58% of the data has the country destination as 'NDF' (No Destination Found). Also, it was found that there is a 1-to-1 correlation betweeen NaN 'date_first_booking' and 'NDF'. That is, the date_first_booking is NaN for all the instances whose country destination is 'NDF'. The high score seen in the above section can be attributed to this factor. So, inorder to get a more realistic model, in this section I will try to build models on the preprocessed data without the date_first_booking information.

### Without setting the max-depth

data_nodfb = data.drop(['dfb_year', 'dfb_month', 'dfb_day'], axis=1)
data_nodfb = pd.DataFrame(preprocessing.StandardScaler().fit_transform(data_nodfb))
foldnum=0
results = pd.DataFrame()
for train, test in cross_validation.KFold(len(data_nodfb), shuffle=True, n_folds=10,
                                           random_state=20160217):
    foldnum+=1
    [tr_data, te_data, tr_labels, te_labels] = folds_to_split(data_nodfb, labels, train, test)
    rf = RandomForestClassifier(n_estimators=600,criterion='gini')
    rf = rf.fit(tr_data, tr_labels.values.ravel())
    rf_predict = rf.predict_proba(te_data) 
    score = ndcg_score(te_labels.as_matrix(), rf_predict, k=5)
    results.loc[foldnum, 'no max-depth + dfb' ] = score

results.mean()

### Varying the max-depth of the Random Forest

results = pd.DataFrame()
foldnum=0
for train, test in cross_validation.KFold(len(data_nodfb), shuffle=True, n_folds=10,
                                           random_state=20160217):
    foldnum+=1
    [tr_data, te_data, tr_labels, te_labels] = folds_to_split(data_nodfb, labels, train, test)

    for max_depth in range(3, 8):
        rf = RandomForestClassifier(n_estimators=600,criterion='gini', max_depth=max_depth)
        rf = rf.fit(tr_data, tr_labels.values.ravel())
        rf_predict = rf.predict_proba(te_data) 
        score = ndcg_score(te_labels.as_matrix(), rf_predict, k=5)
        print 'Fold : {}, Max-depth : {}, Score : {}'.format( foldnum, max_depth, score )
        results.loc[foldnum, 'max_depth=%d'%max_depth ] = score

results.mean()

It can be seen that the nDCG score increases with the max-depth.

### Varying min_samples_leaf

results = pd.DataFrame()
foldnum=0
for train, test in cross_validation.KFold(len(data_nodfb), shuffle=True, n_folds=10,
                                           random_state=20160217):
    foldnum+=1
    [tr_data, te_data, tr_labels, te_labels] = folds_to_split(data_nodfb, labels, train, test)

    for min_sample_leafs in [5, 20, 50]:
        rf = RandomForestClassifier(n_estimators=600,criterion='gini', min_samples_leaf=min_sample_leafs)
        rf = rf.fit(tr_data, tr_labels.values.ravel())
        rf_predict = rf.predict_proba(te_data) 
        score = ndcg_score(te_labels.as_matrix(), rf_predict, k=5)
        print 'Fold : {}, Min_S_Leafs : {}, Score : {}'.format( foldnum, min_sample_leafs, score )
        results.loc[foldnum, 'min_s_leafs=%d'%min_sample_leafs ] = score

results.mean()

The min_sample_leafs 20 and 50 gives almost the same nDCG score.

### Varying max_leaf_nodes

results = pd.DataFrame()
foldnum=0
for train, test in cross_validation.KFold(len(data_nodfb), shuffle=True, n_folds=10,
                                           random_state=20160217):
    foldnum+=1
    [tr_data, te_data, tr_labels, te_labels] = folds_to_split(data_nodfb, labels, train, test)

    for max_leaf_nodes in range(5, 12):
        rf = RandomForestClassifier(n_estimators=600, criterion='gini', max_depth=7, 
                                    max_leaf_nodes=max_leaf_nodes )
        rf = rf.fit(tr_data, tr_labels.values.ravel())
        rf_predict = rf.predict_proba(te_data) 
        score = ndcg_score(te_labels.as_matrix(), rf_predict, k=5)
        print 'Fold : {}, max_leaf_nodes : {}, Score : {}'.format( foldnum, max_leaf_nodes, score )
        results.loc[foldnum, 'max_leaf_nodes=%d'%max_leaf_nodes ] = score

results.mean()

Varying the max_leaf_nodes made the nDCG score worse compared to min_sample_leafs and max_depth values. 

Conclusion :

The best nDCG score among all thest models is 0.823337, which was obtained with min_samples_leaf = 50
