==> 7.Using_session_data.ipynb <==
## Feature Engineering from sessions.csv

import pandas as pd
import numpy as np

from scipy.sparse import coo_matrix, hstack, vstack, csr_matrix

from os import makedirs
from os.path import exists, join

%matplotlib inline

intermediate_path = './intermediate/'

df_s = pd.read_csv('./input/sessions.csv')

df_s.shape

# just checking a random user
#df_s[df_s.user_id == 'awen0gt0iw']

# eliminate user_id with NaN value
print df_s[pd.isnull(df_s.user_id)].shape

df_s = df_s[pd.isnull(df_s.user_id) == False]

# fill NaN for other columns

df_s['action'].fillna('Nada', inplace=True)
df_s['action_detail'].fillna('Nada', inplace=True)
df_s['action_type'].fillna('Nada', inplace=True)

df_s['secs_elapsed'].fillna(0, inplace=True)

# just a check on values that allude to translations
if False:
    df_s = pd.read_csv('./input/sessions.csv')
    l = df_s.action.value_counts().index
    trans = [s for s in l if 'transl' in s ]
    print trans
    l = df_s.action_detail.value_counts().index
    print [s for s in l if 'transl' in s ]
    print df_s[df_s.action.isin(trans)].count()

df_s.shape

df_s.head()

# remove hiphens from action columns.
# '-' are considered separators in CountVectorizer

for c in ['action', 'action_type', 'action_detail']:
    l = df_s[c].unique()
    l = [s for s in l if not pd.isnull(s)]
    print [s for s in l if '-' in s]
    df_s[c] = [ s.replace('-', '') for s in df_s[c]]
    l = df_s[c].unique()
    l = [s for s in l if not pd.isnull(s)]
    print [s for s in l if '-' in s]

# remove spaces from device_type

df_s['device_type'] = [ s.replace(' ', "_") for s in df_s.device_type]

# adding spaces for CountVectorizer split

df_s['device_type']   = ["%s "%w for w in df_s['device_type']]

df_s['action_type']   = ["%s "%w for w in df_s['action_type']]

df_s['action']        = ["%s "%w for w in df_s['action']]
df_s['action_detail'] = ["%s "%w for w in df_s['action_detail']]

# converting secs_elapsed to int 
df_s['secs_elapsed'] = df_s.secs_elapsed.astype(int)

# sessionize action sequence, create session_id column

nw_se_th = 30   # nw_se_th : new session threshold

def generate_session_id(df, th):  

    s_id = 1
    s_ve = []
    
    for se in df['secs_elapsed']:
        if se < th * 60:
            s_ve.append(s_id)
        else:
            s_ve.append(s_id)
            s_id += 1
    df['session_id' + '_' + str(th)] = s_ve
    return df

df_s.secs_elapsed.describe()

# 'sessionizing' user navigation
df_s = df_s.groupby('user_id').apply(lambda x: generate_session_id(x, 30))

# checking a random user
df_s[df_s.user_id == 'd1mm9tcy42']
#df_s[df_s.user_id == 'awen0gt0iw']

# grouping by user_id
# actions and device_type as sum of strings values - concatenation of actions separated by spaces
# secs_elapsed as sum of secs_elapsed int values - total cumulative seconds 
# session_id max- total number of sessions per user

grouped = df_s.groupby('user_id', sort = False)[['action', 'action_type', 'action_detail']]

df_s_a = grouped.sum()

df_s_se = df_s.groupby('user_id', sort = False)['secs_elapsed'].agg({'sum': np.sum})

# selecting columns with session_id counts
session_id_cols = [c for c in df_s.columns if 'session_id' in c]
df_s_sc = df_s.groupby('user_id', sort = False)[session_id_cols].max()

# creating device type datagram
df_s_dt = df_s.groupby('user_id', sort = False)['device_type'].sum()

df_s_se.head()

print df_s_a.shape

print df_s_se.shape

print df_s_sc.shape

print df_s_dt.shape

df_s_a.head()

df_s_a.shape

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer

vec_dt = CountVectorizer(binary=True)

vec_aa = CountVectorizer(ngram_range=(1, 2))
vec_at = CountVectorizer(ngram_range=(1, 2))
vec_ad = CountVectorizer(ngram_range=(1, 2))


df_s_a.shape

# need labels to select train data
df_train = pd.read_csv('./input/train_users_2.csv', index_col=0)

# ids in session file corresponding to train
session_train_id = df_s_a.index.intersection(df_train.index)

session_train_id.shape

df_s_a.shape

# vectorize action, action_type, action_detail all unigrams, all bigrams

vec_dt    = vec_dt.fit(df_s_dt.ix[session_train_id].values)
vec_aa    = vec_aa.fit(df_s_a.ix[session_train_id].action.values)
vec_at    = vec_at.fit(df_s_a.ix[session_train_id].action_type.values)
vec_ad    = vec_ad.fit(df_s_a.ix[session_train_id].action_detail.values)

#session_all_aa_ad = vec_aa_ad.transform(df_s_a.aa_ad.values)

session_all_dt    = vec_dt.transform(df_s_dt.values)
session_all_aa    = vec_aa.transform(df_s_a.action.values)
session_all_at    = vec_at.transform(df_s_a.action_type.values)
session_all_ad    = vec_ad.transform(df_s_a.action_detail.values)

# checking vectorized matrices shapes
print session_all_dt.shape
print session_all_aa.shape
print session_all_at.shape
print session_all_ad.shape

# concatenating (action, action_detail) with action_type 

#session_all_a = hstack([session_all_aa_ad, session_all_at])
session_all_a = hstack([session_all_at, session_all_aa, session_all_ad])

session_all_a = session_all_a.tocsr()

#### Building session_all

# add column 'unique_devices' number of unique devices
session_all = pd.DataFrame(session_all_dt.sum(axis = 1), index=df_s_se.index, columns=['unique_devices'])

# add columns for vectorized device types 
#session_all = pd.DataFrame(session_all_dt.toarray(), index=df_s_a.index, columns=vec_dt.get_feature_names())

# add column 'secs elapsed' cumulative seconds elapsed
session_all = pd.concat([session_all, df_s_se], axis = 1)
print session_all.shape

# add column 'session_id' represents total number of sessions
session_all = pd.concat([session_all, df_s_sc], axis = 1)
#session_all.rename(columns={'session_id' : 'session_count'}, inplace=True)

session_all.head()

# Univariate feature selection for action uni and bigrams

# Creating and Encoding y
from sklearn.preprocessing import LabelEncoder
labels = df_train.ix[session_train_id].country_destination
le = LabelEncoder()
y = le.fit_transform(labels.values) 

from sklearn.feature_selection import SelectKBest, chi2

X = session_all_a[:session_train_id.shape[0]].copy()
sel = SelectKBest(chi2, k=1000)
sel.fit(X,y)

X_new = sel.transform(session_all_a)

X_new = pd.DataFrame(X_new.toarray(), index=df_s_a.index)

# adding action uni and bigram features after feature selection

session_all = pd.concat([session_all, X_new], axis = 1)

session_all.shape

session_all.head()

if not exists(intermediate_path):
    makedirs(intermediate_path)

session_all.to_csv( join(intermediate_path, 'session_features_1000.csv') )
