==> 8.Final_notebook.ipynb <==

Single XGB Classifier after feature engineering

# Basic ML modules
import numpy as np
import pandas as pd
from sklearn.preprocessing import LabelEncoder, StandardScaler
from xgboost.sklearn import XGBClassifier
from sklearn.ensemble import GradientBoostingClassifier, RandomForestClassifier, AdaBoostClassifier
from sklearn.cross_validation import cross_val_score, KFold

from time import time
from datetime import datetime, timedelta, date

# Calendar modules - needed for holidays
from workalendar.usa import UnitedStates
from workalendar.europe import France, Germany, Italy, UnitedKingdom, Portugal
from workalendar.oceania import Australia

# Modules for plots
import matplotlib as mpl
import matplotlib.pyplot as plt
import seaborn as sns

# Command so that plots appear in the iPython Notebook
%matplotlib inline

# File management modules
from os import makedirs
from os.path import exists, join

np.random.seed(0)

intermediate_path = './intermediate/'
output_path = './output/'

# Load data
df_train = pd.read_csv('./input/train_users_2.csv')
df_test = pd.read_csv('./input/test_users.csv')


# Create labels 
labels = df_train['country_destination'].values
df_train = df_train.drop(['country_destination'], axis=1)
id_test = df_test['id']
piv_train = df_train.shape[0]

df_train.head()

df_test.head()

df_train.shape

# Creating and Encoding y
le_y = LabelEncoder()
y = le_y.fit_transform(labels) 

# Creating a DataFrame with train+test data
df_all = pd.concat((df_train, df_test), axis=0, ignore_index=True)

# dropping column not available in test data
df_all = df_all.drop('date_first_booking', axis=1)

df_all.head()

# Reformatting columns with time data and some feature engineering

# from string to datetime format
df_all['datehour_timestamp_first_active'] = df_all['timestamp_first_active'].astype(str)
df_all['date_timestamp_first_active'] = [ d[:8] for d in df_all['datehour_timestamp_first_active'] ]
df_all['hour_timestamp_first_active'] = [ d[8:] for d in df_all['datehour_timestamp_first_active'] ]

df_all['date_timestamp_first_active'] = [ datetime.strptime(d, '%Y%m%d') for d in df_all['date_timestamp_first_active']]
df_all['hour_timestamp_first_active'] = [ datetime.strptime(h, '%H%M%S') for h in df_all['hour_timestamp_first_active']]

df_all['date_account_created'] = [ datetime.strptime(d, '%Y-%m-%d') for d in df_all['date_account_created']]

 
# create d_ac-fa: number of days between account creation and first access
df_all['d_ac-fa'] = df_all['date_account_created'] - df_all['date_timestamp_first_active']
df_all['d_ac-fa'] = df_all['d_ac-fa'].astype('int')/8.64e13  #  CHECK!!!!

# create mo_first_active: month of the first active
df_all['mo_first_active'] = [ d.month if not pd.isnull(d) else 0 for d in df_all['date_timestamp_first_active'] ]
df_all['ho_first_active'] = [ d.hour if not pd.isnull(d) else 0 for d in df_all['hour_timestamp_first_active'] ]

# create ye_first_active: year of the first active
# NOT USED
#df_all['ye_first_active'] = [ d.year if not pd.isnull(d) else 0 for d in df_all['date_timestamp_first_active'] ]

# create weekday_first_active: day of week (categorical variable)
df_all['weekday_first_active'] = [ datetime.strftime(d,'%a') if not pd.isnull(d) else d for d in df_all['date_timestamp_first_active']]

# create features
# d_ac : day of the month account creation 
# d_fa : day of the month first access
df_all['d_ac'] = [ d.day if not pd.isnull(d) else 0 for d in df_all['date_account_created']  ]    
df_all['d_fa'] = [ d.day if not pd.isnull(d) else 0 for d in df_all['date_timestamp_first_active']  ]

# Transform timestamp into date format
df_all['date_timestamp_first_active'] = [ d.date() for d in df_all.date_timestamp_first_active]

df_all.head()

# checkpoint 0
df_all_ck0 = df_all.copy()

# revert to checkpoint 0
#df_all = df_all_ck0.copy()

### Feature Engineering: Including Holidays

def daterange(start_date, end_date):
    for n in range(int ((end_date - start_date).days)):
        yield start_date + timedelta(n)

# create days_to_next_CC_hol 
# CC is the country code
# 

country_dict = { 'NDF': 'None', 'US': UnitedStates(), 'other': 'None', 'FR': France(), 
 'IT': Italy(), 'GB': UnitedKingdom(), 'ES': 'None', 'CA': 'None', 'DE': Germany(), 
 'NL' : 'None', 'AU' : Australia(), 'PT' : Portugal() }

# only end up using holidays in USA
country_dict = { 'NDF': 'None', 'US': UnitedStates(), 'other': 'None', 'FR': 'None', 
 'IT': 'None', 'GB': 'None', 'ES': 'None', 'CA': 'None', 'DE': 'None', 
 'NL' : 'None', 'AU' : 'None', 'PT' : 'None' }

for c in country_dict:
    if country_dict[c] != 'None':
        hol = country_dict[c].holidays(2008)
        for year in range(2009, 2016, 1):
            hol += country_dict[c].holidays(year)
        
        hol = pd.Series([h[0] for h in hol ])
        
        hol_dict = {}
        for d in daterange( date(2009, 1, 1), date(2014, 12, 31) ):
            delta_list = hol - d
            delta_list = [dd.days for dd in delta_list]
            next_holiday_delta = min([n for n in delta_list if n >= 0])
            nex2_holiday_delta = min([n for n in delta_list if n > next_holiday_delta])
            hol_dict[d] = (next_holiday_delta, nex2_holiday_delta)
            
            column_name = 'days_to_next_' + c + '_hol'
        df_all[column_name] = [ hol_dict[d][0] for d in df_all['date_timestamp_first_active'] ]

df_all.shape

# creating country ranking columns

for c in country_list:
    df_all[c + '_rank'] = [ c_am[a, g][c]  if (not pd.isnull(a)) & (a < 105) else -999
                           for (a, g) in df_all[['age', 'gender']].itertuples(index=False) ]

df_all.head()

# removing features not going to be used
features = list(df_all.columns)

remove = ['date_account_created',
 'timestamp_first_active',
 'datehour_timestamp_first_active',
 'date_timestamp_first_active',
 'hour_timestamp_first_active']

for rf in remove:
    features.remove(rf)

df_all = df_all[features]

df_all.head()

df_all.shape

# checking age distribution
if False:
    sns.set_context("talk")
    box = sns.distplot(df_all.age.dropna(), hist_kws={"range" : (0, 100)}, bins = 50)
    box.set(xlim=(0, 100))

# checkpoint 1
df_all_ck1 = df_all.copy()

# revert to checkpoint
#df_all = df_all_ck1.copy()

# Note age has a lot of NaN
# and first_affiliate_tracked has few
df_all.count()

# imputation -1 for ages NaN
df_all['age'].fillna(-999, inplace=True)

# imputation most frequent 'untracked' for first_affiliate_tracked
df_all['first_affiliate_tracked'].fillna(df_all.first_affiliate_tracked.value_counts().index[0], inplace=True)

# checkpoint 2
df_all_ck2 = df_all.copy()

# revert to checkpoint
#df_all = df_all_ck2.copy()

df_all.head()

# Remove age outliers

av = df_all.age.values
df_all['age'] = np.where(np.logical_or(av<14, av>104), -999, av)

# reduce cardinality of feature
df_all['days_to_next_US_hol'] = np.ceil((df_all.days_to_next_US_hol / 2))

df_all.head()

# removing features based on feature_importances_
features = list(df_all.columns)

remove = ['d_ac-fa','signup_app', 'd_ac']

for rf in remove:
    features.remove(rf)

df_all = df_all[features]

# checking any NaN value left
df_all.info()

# building a list of categorical variables to be LabelEncoded 
to_le = [ i for i, t in zip(df_all.dtypes.index, df_all.dtypes.values)  if t == 'O']
to_le.remove('id')

# LabelEncode variables

le = LabelEncoder()
for ri in to_le: # ri: replaced by integer
    le.fit(df_all[ri])   # takes labels and assigns each to a integer,
    df_all[ri] = le.transform(df_all[ri])
    print le.classes_
    print le.transform(le.classes_)
    print "------"
    print

df_all.head()

print y.shape
print df_all.shape

# checkpoint 3
df_all_ck3 = df_all.copy()

# revert to checkpoint 3
df_all = df_all_ck3.copy()

### Merge with session data

# Load session features processed from sessions.csv in
# Extracting features from sessions.csv

df_s = pd.read_csv(join(intermediate_path, 'session_features_1000.csv'))
df_s = df_s.rename(columns = {'user_id': 'id'})

df_s.head()

# merging 
# df_all (original users data) with 
# df_s (processed session data)
df_all = pd.merge(df_all, df_s, how = 'left', on = 'id')

df_all.head()

df_all = df_all.drop('id', axis=1)

df_all.shape

# filling NaN with -1
df_all.fillna(-999, inplace = True)

df_all.head()

#Splitting train and test
vals = df_all.values
X = vals[:piv_train]

X_test = vals[piv_train:]

# scaling features
from sklearn.preprocessing import MinMaxScaler, StandardScaler

#scaler = MinMaxScaler()
scaler = StandardScaler()
X = scaler.fit_transform(X)
X_test = scaler.transform(X_test)

## Learning and predicting, and generating submission file



# Learning and predicting 
# Generating submission file

clf = XGBClassifier(max_depth=6, learning_rate=0.2, n_estimators=25, #reg_lambda=3.33,
                    objective='multi:softprob', subsample=0.75, colsample_bytree=0.5, seed=0)

# Fit and predict
clf.fit(X, y)
y_pred = clf.predict_proba(X_test)  

# Building submission files

# Taking the 5 classes with highest probabilities
ids = []  #list of ids
cts = []  #list of countries
for i in range(len(id_test)):
    idx = id_test[i]
    ids += [idx] * 5
    cts += le_y.inverse_transform(np.argsort(y_pred[i])[::-1])[:5].tolist()

#Generate submission
if not exists(output_path):
    makedirs(output_path)

sub = pd.DataFrame(np.column_stack((ids, cts)), columns=['id', 'country'])
f = 'sub.' + datetime.now().strftime("%d%m%y_%H%M%S") + '.csv'
sub.to_csv(join(output_path, f), index=False)

### Cross-validation

def scorer(clf, X_trn, y_trn, X_val, y_val):
    clf.fit(X_trn, y_trn)
    y_pred = clf.predict_proba(X_val)  
    
    cts = []  #list of countries
    for i in range(len(y_val)):
        cts.append( np.argsort(y_pred[i])[::-1][:5].tolist() )

    cts = pd.DataFrame(cts)
    y_val = pd.Series(y_val)
    return score_predictions(cts, y_val)



def dcg_at_k(r, k, method=1):
    r = np.asfarray(r)[:k]
    if r.size:
        if method == 0:
            return r[0] + np.sum(r[1:] / np.log2(np.arange(2, r.size + 1)))
        elif method == 1:
            return np.sum(r / np.log2(np.arange(2, r.size + 2)))
        else:
            raise ValueError('method must be 0 or 1.')
    return 0.


def ndcg_at_k(r, k=5, method=1):
    dcg_max = dcg_at_k(sorted(r, reverse=True), k, method)
    if not dcg_max:
        return 0.
    return dcg_at_k(r, k, method) / dcg_max


def score_predictions(preds, truth, n_modes=5):
    """
    preds: pd.DataFrame
      one row for each observation, one column for each prediction.
      Columns are sorted from left to right descending in order of likelihood.
    truth: pd.Series
      one row for each obeservation.
    """
    assert(len(preds)==len(truth))
    r = pd.DataFrame(0, index=preds.index, columns=preds.columns, dtype=np.float64)
    for col in preds.columns:
        r[col] = (preds[col] == truth) * 1.0

    score = pd.Series(r.apply(ndcg_at_k, axis=1, reduce=True), name='score')
    return score
    
    
    
preds = pd.DataFrame([['US','FR'],['FR','US'],['FR','FR']])
truth = pd.Series(['US','US','FR'])
print'predictions:' 
print preds
print'truth:'
print truth
print'scores:'
print score_predictions(preds, truth)

i3 = 87628
i4 = 113088
i5 = 170158
i6 = len(X)

# fold1: same quarter one year before
# fold2: one quarter before the test data

fold1_train = np.array(range(0, i3) + range(i4, i6))
fold1_test  = np.array(range(i3, i4))
fold2_train = np.array(range(0, i5))
fold2_test  = np.array(range(i5, i6))

n_folds = 2
kf = [(fold1_train, fold1_test), (fold2_train, fold2_test)]




means = {}

for n_estimators in [23]:
    for subsample in [0.75]:
        for learning_rate in [0.2]:
            t0 = time()
            
            clf = XGBClassifier(max_depth=6, learning_rate=learning_rate, n_estimators=n_estimators, reg_lambda=10,
                        objective='multi:softprob', subsample=subsample, colsample_bytree=0.5, seed=0)  
            
                    
            ndcg_scores = []
            
            for train, test in kf:
                cv_pred = scorer(clf, X[train], y[train], X[test], y[test])
                ndcg_scores.append( np.mean(cv_pred) )
            
            print
            print "learning rate:", learning_rate
            print "n_estimators:", n_estimators
            print "subsample:", subsample
            print
            print n_folds, "folds NDCG scores: ", [ "%.4f" % s for s in ndcg_scores]
            print "mean:", np.mean(ndcg_scores)
            means[(n_estimators, subsample, learning_rate)] = ndcg_scores 
            print
            print "Classifier:"
            print clf
            print
            print "Running time:", "%.2f" % (time() - t0)
            print "******************************************************************************"

max(means, key = lambda x: means.get(x) )

means

df_means['average_ndcg'] = df_means.mean(axis=1)
df_means

df_means = pd.DataFrame(means).T
df_means['average_ndcg'] = df_means.mean()
df_means['average_ndcg'] = df_means.mean(axis=1)
df_means

df_means = pd.DataFrame(means).T
df_means['average_ndcg'] = df_means.mean()
df_means['average_ndcg'] = df_means.mean(axis=1)
df_means

df_means = pd.DataFrame(means).T
df_means['average_ndcg'] = df_means.mean()
df_means['average_ndcg'] = df_means.mean(axis=1)
df_means

df_means = pd.DataFrame(means).T
df_means['average_ndcg'] = df_means.mean()
df_means['average_ndcg'] = df_means.mean(axis=1)
df_means

# FS 1000
# lambda_reg = 10
df_means = pd.DataFrame(means).T
df_means['average_ndcg'] = df_means.mean()
df_means['average_ndcg'] = df_means.mean(axis=1)
df_means

The scripts give a NDCG score of about 0.8845 , On the kaggle submissions . which is around 140 position in this competition .
